REF: implement test_tz_localize
